+++ src/types.rs
@@ -1,7 +1,6 @@
 use std::cell::RefCell;
 use std::rc::Rc;
 
-use crate::controller::Controller;
 use crate::graphics::ppu::Ppu;
 use crate::interfaces::Memory;
 use crate::processor::bus::Bus;
@@ -16,5 +15,3 @@ pub type SharedCiram = Rc<RefCell<Ciram>>;
 pub type SharedMirroredRom = Rc<RefCell<MirroredMemory<Rom>>>;
 
 pub type SharedPpu = Rc<RefCell<Ppu>>;
-
-pub type SharedController = Rc<RefCell<Controller>>;